# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: credo-fabric-ai
  version: 0.2.0
  summary: Analytics API over Microsoft Fabric for the CREDICORP executive dashboard
  description: |
    Read-only analytics API over `Credicorp_lakehouse`. It backs the CREDICORP executive
    dashboard: KPI tiles, value aggregations, period-over-period deltas, breakdowns, trends,
    drill-down tables, and a natural-language endpoint.

    ## Two endpoints do most of the work

    - **`GET /dashboard`** returns a *renderable layout* — for every dataset the caller can
      read, which columns are measures, which are chartable dimensions, which is the time
      axis, and the exact endpoint each tile should call. The frontend loops over tiles; it
      never hardcodes a table or column name.
    - **`GET /dashboard/pulse`** returns the *cross-platform rollup* — total records, how many
      sit in a state needing action, total financial exposure, and period movement, across
      every system the caller can see.

    ## Schema-driven

    Table and column names are validated against the live schema on every request. Nothing is
    hardcoded, so the API works against the lakehouse without its columns being documented
    here — and that validation is the injection boundary, since SQL identifiers cannot be
    bound as parameters. Values are always bound.

    ## Roles

    The caller's role arrives in the `x-role` header (or `?role=`) and decides which datasets
...
    the database.

    | Role | Datasets |
...

...

...

...

...

...
servers:
...
paths:
...
  /capabilities:
    get:
      tags: [Service]
      summary: Aggregations, buckets and filter operators this build supports
      description: |
        Lets the frontend build filter and metric pickers from the server's actual capability
        list instead of a hardcoded copy that drifts.
      operationId: getCapabilities
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      responses:
        '200':
          description: Supported grammar.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Capabilities' }
  /roles:
    get:
      tags: [Service]
      summary: Valid roles and the one currently active
      operationId: getRoles
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      responses:
        '200':
          description: Role list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  roles: { type: array, items: { $ref: '#/components/schemas/Role' } }
                  activeRole: { $ref: '#/components/schemas/Role' }
                  note: { type: string, examples: ['Prototype: role is unverified.'] }

  # ----------------------------------------------------------------- Schema --
  /schema:
    get:
      tags: [Schema]
      summary: Every dataset this role can read
      description: |
        Includes each dataset's time axis and its numeric measures, already filtered to the
        columns worth aggregating (ids and reference numbers are excluded).
      operationId: getSchema
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      responses:
        '200':
          description: Datasets visible to the role.
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      role: { $ref: '#/components/schemas/Role' }
                      tables:
                        type: array
                        items: { $ref: '#/components/schemas/TableSchema' }
                  - $ref: '#/components/schemas/SyntheticMarker'
  /schema/{table}:
    get:
      tags: [Schema]
      summary: One dataset's columns
      operationId: getTableSchema
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
        - $ref: '#/components/parameters/Table'
      responses:
        '200':
          description: Column list.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/TableSchema'
                  - $ref: '#/components/schemas/SyntheticMarker'
        '403': { $ref: '#/components/responses/Forbidden' }
  /schema/{table}/values:
    get:
      tags: [Schema]
      summary: Distinct values in a column
      description: Populates filter dropdowns. Returns each value with its row count, highest first.
      operationId: getColumnValues
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
        - $ref: '#/components/parameters/Table'
        - name: column
          in: query
          required: true
          description: Column to enumerate.
          schema: { type: string }
          example: status
        - name: limit
          in: query
          description: Maximum distinct values (1–200).
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        '200':
          description: Distinct values with counts.
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      table: { type: string }
                      column: { type: string }
                      values:
                        type: array
                        items:
                          type: object
                          properties:
                            value: { type: [string, number, boolean, 'null'] }
                            count: { type: integer }
                  - $ref: '#/components/schemas/SyntheticMarker'
              example:
                table: vendor_registry
                column: status
                values:
                  - { value: Active, count: 396 }
                  - { value: Under Review, count: 99 }
                  - { value: Pending Onboarding, count: 64 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # -------------------------------------------------------------- Dashboard --
  /dashboard:
    get:
      tags: [Dashboard]
      summary: Renderable tile and chart layout for this role
      description: |
        The layout contract. For each accessible dataset it returns a section containing:

        - **`tiles`** — KPI tiles. A row count, plus up to three measure tiles. Money-like
          columns get `fn=sum` and `format: currency`; other measures get `fn=avg`.
        - **`charts`** — a time series on the dataset's time axis, plus breakdowns for text
          columns with few enough distinct values to read as a chart. `type` is `donut` at
          six values or fewer, `bar` above that.
        - **`table`** — the drill-down grid, with a display label and kind per column.

        Every entry carries the `endpoint` to call for its data, already assembled with query
        parameters. Tiles that support a trend arrow also carry `trendEndpoint`.

        Because it is derived from the live schema, a new table appearing in Fabric shows up
        on the dashboard with no code change.
      operationId: getDashboardSpec
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
      responses:
        '200':
          description: Layout spec.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DashboardSpec'
                  - $ref: '#/components/schemas/SyntheticMarker'
  /dashboard/pulse:
    get:
      tags: [Dashboard]
      summary: Cross-dataset rollup — totals, pressure and exposure
      description: |
        One number per question an executive actually asks, spanning every dataset the role
        can see rather than describing one:

        - **`headline.needsAttention`** — records in a state that means someone still has to
          act (pending, open, escalated, suspended, denied, critical…). Closed states such as
          *Rejected* and *Terminated* are deliberately excluded, since counting them inflates
          the figure.
        - **`headline.financialExposure`** — the sum of each dataset's most money-like measure.
        - **`attention`** / **`exposure`** — the same, broken out per dataset and pre-sorted,
          so the UI can render "where the pressure is" without sorting client-side.
      operationId: getPulse
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
        - name: days
          in: query
          description: Comparison window for the period-over-period figures.
          schema: { type: integer, minimum: 1, maximum: 3650, default: 30 }
      responses:
        '200':
          description: Organisation rollup.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Pulse'
                  - $ref: '#/components/schemas/SyntheticMarker'

  # ---------------------------------------------------------------- Metrics --
...
  /data/{table}:
    get:
      tags: [Data]
      summary: Paginated drill-down rows
      description: |
        Returns whole rows. Sorting defaults to the dataset's time axis descending; pagination
        is capped at 500 rows per page.

        Any query parameter that is not a reserved key is treated as a column filter, using the
        same grammar as everywhere else.
      operationId: getRows
      parameters:
        - $ref: '#/components/parameters/RoleHeader'
        - $ref: '#/components/parameters/Table'
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          description: Silently clamped to 500.
          schema: { type: integer, minimum: 1, maximum: 500, default: 50 }
        - name: sort
          in: query
          description: Column to sort by. Defaults to the dataset's time axis.
          schema: { type: string }
        - name: order
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/DateColumn'
        - $ref: '#/components/parameters/Filters'
      responses:
        '200':
          description: A page of rows.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/DataPage'
                  - $ref: '#/components/schemas/SyntheticMarker'
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # --------------------------------------------------------------------- AI --
  /ai/ask:
    post:
      tags: [AI]
      summary: Answer a question in natural language
      description: |
        The model never writes SQL. It selects from a fixed set of role-gated tools
        (`list_datasets`, `organisation_pulse`, `count_rows`, `aggregate_metric`,
        `compare_periods`, `breakdown_by_column`, `trend_over_time`, `sample_records`) and each
        one checks the caller's role before touching data. A denied request therefore never
        reaches the database, and a hallucinated table name fails schema resolution rather
        than executing.

        The `trace` shows which tools ran with what arguments, so the dashboard can display
        provenance instead of an unsourced figure. A tool that was refused has `denied: true` —
        surface that rather than hiding it.

        Requires Azure OpenAI to be configured; returns `503` if it is not.
      operationId: askQuestion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question:
                  type: string
                  description: The question, in plain language.
                  examples: ['what is our total travel spend by department this year?']
                role:
                  $ref: '#/components/schemas/Role'
            examples:
              aggregate:
                summary: A value question
                value: { question: 'what is our total travel spend by department this year?', role: admin }
              summary:
                summary: A broad question — resolves via the pulse tool
                value: { question: 'give me an executive summary of what needs attention right now', role: admin }
      responses:
        '200':
          description: |
            The answer and its provenance. Note this is `200` even when the model could not
            answer or every tool was denied — read `trace` to distinguish those cases.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/AskResponse'
                  - $ref: '#/components/schemas/SyntheticMarker'
              examples:
                answered:
                  summary: Answered from a tool result
                  value:
                    answer: There are 486 memos currently pending approval.
                    role: manager
                    trace:
                      - tool: count_rows
                        args: { table: credo_memos, filters: { status: Pending Approval } }
                        denied: false
                        ok: true
                denied:
                  summary: Role could not see the data
                  value:
                    answer: There is no dataset available related to memos. I cannot answer this question.
                    role: staff
                    trace: []
        '400': { $ref: '#/components/responses/BadRequest' }
        '503':
          description: Azure OpenAI is not configured on this deployment.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ------------------------------------------------------------------ Cache --
  /cache/invalidate:
    post:
      tags: [Cache]
      summary: Drop cached results
      description: |
        Omit `prefix` to clear everything, or pass one to clear a slice — cache keys are
        prefixed by kind: `count:`, `breakdown:`, `timeseries:`, `aggregate:`, `compare:`,
        `distinct:`, `profile:`, `schema:`.

        **Unauthenticated write endpoint** — put it behind auth before deployment.
      operationId: invalidateCache
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                prefix:
                  type: string
                  description: Key prefix to clear. Omit to clear the whole cache.
                  examples: ['breakdown:credo_memos']
      responses:
        '200':
          description: Cache state after the drop.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  cache: { $ref: '#/components/schemas/CacheStats' }

# ============================================================== Components ==
components:
...
  securitySchemes:
    RoleHeader:
      type: apiKey
      in: header
      name: x-role
      description: |
        **Prototype only, and not a security mechanism.** The role is read from this header
        with no verification whatsoever, so any caller can claim any role. It is modelled as a
        security scheme purely so Swagger UI offers a box to set it.

        Replace with a validated token claim before this is exposed beyond the demo.

  # ------------------------------------------------------------ Parameters --
  parameters:
    RoleHeader:
      name: x-role
      in: header
      required: false
      description: |
        Caller's role. Falls back to `DEFAULT_ROLE` (`admin` in development) when absent.
        An unrecognised value also falls back — it never escalates.
      schema: { $ref: '#/components/schemas/Role' }
    Table:
      name: table
      in: path
      required: true
      description: |
        Dataset name, case-insensitive. Validated against the live schema — an unknown name is
        a `404`, and a name the role cannot read is a `403` regardless of whether it exists.
      schema: { type: string }
      examples:
        tickets: { value: credo_support_tickets }
        vendors: { value: vendor_registry }
        memos: { value: credo_memos }
        travel: { value: credo_travel_details }
        audit: { value: craig_audit_logs }
    From:
      name: from
      in: query
      description: |
        Start of a date range, applied to the dataset's time axis **without the caller needing
        to know its column name**. Override the column with `dateColumn`.
      schema: { type: string, format: date }
      example: '2026-01-01'
    To:
      name: to
      in: query
      description: End of the date range, inclusive.
      schema: { type: string, format: date }
      example: '2026-03-31'
    DateColumn:
      name: dateColumn
      in: query
      description: Which temporal column `from`/`to` apply to. Defaults to the dataset's time axis.
      schema: { type: string }
      example: created_at
    Filters:
      name: filters
      in: query
      style: form
      explode: true
      required: false
      description: |
        **Free-form column filters.** Any query parameter that is not a reserved key is read as
        a filter on the column of that name. A bare value means equality; anything richer takes
        an operator prefix.

        | Expression | Meaning |
        | --- | --- |
        | `?status=Open` | `status = 'Open'` |
        | `?status=ne:Draft` | `status <> 'Draft'` |
        | `?status=in:Open,Escalated` | `status IN (…)` |
        | `?status=nin:Draft,Closed` | `status NOT IN (…)` |
        | `?amount_requested=gte:500000` | `>=` — also `gt`, `lt`, `lte` |
        | `?created_at=between:2026-01-01,2026-03-31` | inclusive range |
        | `?vendor_name=contains:Atlantic` | `LIKE '%Atlantic%'` |
        | `?vendor_name=startswith:Atl` | `LIKE 'Atl%'` |
        | `?resolved_at=isnull` | `IS NULL` |
        | `?resolved_at=notnull` | `IS NOT NULL` |

        Values are coerced to the column's type; one that does not fit is a `400`. An operator
        that does not suit the type is also a `400` — `gte` on text, `contains` on a number.
        Repeating a parameter is read as an `in` list.

        Null handling follows SQL: a `NULL` fails every comparison, including `ne`. Use
        `isnull` / `notnull` explicitly.

        Applied filters are echoed back on the response as `filters`, so the UI can render chips.

        **Reserved keys** (never treated as filters): `page`, `pageSize`, `sort`, `order`,
        `role`, `column`, `limit`, `bucket`, `days`, `from`, `to`, `dateColumn`, `fn`,
        `groupBy`, `metrics`.
      schema:
        type: object
        additionalProperties: { type: string }
      examples:
        equality:
          summary: Equality
          value: { status: Approved }
        combined:
          summary: Several filters combine with AND
          value: { status: 'in:Pending Approval,Under Review', amount_requested: 'gte:500000' }
        nullCheck:
          summary: Unresolved records
          value: { resolved_at: isnull }

  # --------------------------------------------------------------- Schemas --
  schemas:
    Role:
      type: string
      enum: [staff, manager, admin]
      description: |
        `staff` sees support tickets and vendors; `manager` adds memos and travel; `admin` adds
        the compliance/AML audit log.
...
    SyntheticMarker:
      type: object
      description: |
        Present on every response **only** when the synthetic data source is active. Both fields
        are absent when data is live, so their presence is the reliable signal.
      properties:
        synthetic:
          type: boolean
          const: true
          description: These figures are not real lakehouse data.
        source:
          type: string
          const: mock
...
    CacheStats:
      type: object
      properties:
        total: { type: integer, description: Entries held, fresh or stale. }
        fresh: { type: integer, description: Entries still within their TTL. }
        inflight: { type: integer, description: Producers running now, deduped by single-flight. }
...
    Capabilities:
      type: object
      properties:
        role: { $ref: '#/components/schemas/Role' }
        aggregations: { type: array, items: { $ref: '#/components/schemas/Aggregation' } }
        buckets: { type: array, items: { $ref: '#/components/schemas/Bucket' } }
        filterOperators:
          type: array
          items: { type: string }
          examples: [[eq, ne, gt, gte, lt, lte, in, nin, between, contains, startswith, isnull, notnull]]
        filterSyntax: { type: string }
        dataSource: { $ref: '#/components/schemas/DataSource' }
...
    TableSchema:
      type: object
      properties:
        table: { type: string }
        schema: { type: string, examples: [dbo] }
        columns: { type: array, items: { $ref: '#/components/schemas/Column' } }
        timeAxis:
          type: [string, 'null']
          description: Column used for trends and `from`/`to` when none is named.
        measures:
          type: array
          items: { type: string }
          description: |
            Numeric columns worth aggregating. Ids and reference numbers are excluded — a tile
            reading "Total ticket_id" would discredit the page.
...
    DataPage:
      type: object
      properties:
        table: { type: string }
        rows:
          type: array
          description: Whole source rows. Shape follows the dataset — see `/schema/{table}`.
          items: { type: object, additionalProperties: true }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer, description: Matching rows before pagination. }
        totalPages: { type: integer }
        sort: { type: string }
        order: { type: string, enum: [ASC, DESC] }
        filters: { type: array, items: { $ref: '#/components/schemas/FilterEcho' } }
...
    DashboardSpec:
      type: object
      properties:
        role: { $ref: '#/components/schemas/Role' }
        generatedAt: { type: string, format: date-time }
        summaryEndpoint: { type: string, examples: ['/api/v1/dashboard/pulse'] }
        sections: { type: array, items: { $ref: '#/components/schemas/DashboardSection' } }
    Pulse:
      type: object
      properties:
        role: { $ref: '#/components/schemas/Role' }
        windowDays: { type: integer }
        generatedAt: { type: string, format: date-time }
        headline:
          type: object
          properties:
            datasets: { type: integer }
            totalRecords: { type: integer }
            newThisPeriod: { type: integer }
            previousPeriod: { type: integer }
            percentChange: { type: [number, 'null'] }
            needsAttention:
              type: integer
              description: Records in a state requiring action. Closed states are excluded.
            financialExposure:
              type: number
              description: Sum of each dataset's most money-like measure.
        attention:
          type: array
          description: Per dataset, sorted by count descending.
          items:
            type: object
            properties:
              dataset: { type: string }
              label: { type: string }
              column: { type: string, description: The status-like column this was read from. }
              count: { type: integer }
              share: { type: number }
              states:
                type: array
                items:
                  type: object
                  properties:
                    label: { type: string }
                    value: { type: integer }
        exposure:
          type: array
          description: Per dataset, sorted by total descending.
          items:
            type: object
            properties:
              dataset: { type: string }
              label: { type: string }
              column: { type: string }
              measureLabel: { type: string }
              total: { type: number }
        datasets:
          type: array
          description: The per-dataset detail the headline was rolled up from.
          items:
            type: object
            properties:
              dataset: { type: string }
              label: { type: string }
              rows: { type: integer }
              trend:
                oneOf:
                  - { $ref: '#/components/schemas/Trend' }
                  - { type: 'null' }
              exposure: { type: ['object', 'null'] }
              attention: { type: ['object', 'null'] }
...
  responses:
    BadRequest:
      description: |
        The request was understood but is not valid — an operator that does not suit the
        column's type, a value that will not coerce, a missing required parameter.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            typeMismatch:
              value: { error: "Operator 'gte' needs a numeric or date column, not 'status'", name: BadRequestError, requestId: r7 }
            badValue:
              value: { error: "'not-a-date' is not a date (column 'created_at')", name: BadRequestError, requestId: r8 }
    Forbidden:
      description: |
        The role may not read this dataset. Returned **before** the table is resolved, so a
        caller cannot distinguish a dataset they cannot see from one that does not exist.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: "Role 'staff' is not permitted to read 'credo_memos'", name: AccessDeniedError, requestId: r9 }
    NotFound:
      description: The column does not exist on this dataset.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: "Unknown column 'nope' on 'credo_memos'", name: UnknownObjectError, requestId: ra }
security:
  - RoleHeader: []